chore: move encoding functionality to new package inside internal utils
