+++ sdk/evm/encoder.go
@@ -7,6 +7,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 
+	abiUtils "github.com/smartcontractkit/mcms/internal/utils/abi"
 	"github.com/smartcontractkit/mcms/sdk"
 	"github.com/smartcontractkit/mcms/sdk/evm/bindings"
 	"github.com/smartcontractkit/mcms/types"
@@ -58,7 +59,7 @@ func (e *EVMEncoder) HashOperation(
 	}
 
 	abi := `[{"type":"bytes32"},{"type":"tuple","components":[{"name":"chainId","type":"uint256"},{"name":"multiSig","type":"address"},{"name":"nonce","type":"uint40"},{"name":"to","type":"address"},{"name":"value","type":"uint256"},{"name":"data","type":"bytes"}]}]`
-	encoded, err := ABIEncode(abi, mcmDomainSeparatorOp, bindOp)
+	encoded, err := abiUtils.ABIEncode(abi, mcmDomainSeparatorOp, bindOp)
 	if err != nil {
 		return common.Hash{}, err
 	}
@@ -70,7 +71,7 @@ func (e *EVMEncoder) HashOperation(
 // ManyChainMultiSig contract, and hashes it.
 func (e *EVMEncoder) HashMetadata(metadata types.ChainMetadata) (common.Hash, error) {
 	abi := `[{"type":"bytes32"},{"type":"tuple","components":[{"name":"chainId","type":"uint256"},{"name":"multiSig","type":"address"},{"name":"preOpCount","type":"uint40"},{"name":"postOpCount","type":"uint40"},{"name":"overridePreviousRoot","type":"bool"}]}]`
-	encoded, err := ABIEncode(abi, mcmDomainSeparatorMetadata, e.ToGethRootMetadata(metadata))
+	encoded, err := abiUtils.ABIEncode(abi, mcmDomainSeparatorMetadata, e.ToGethRootMetadata(metadata))
 	if err != nil {
 		return common.Hash{}, err
 	}
