+++ sdk/evm/utils.go
@@ -1,10 +1,6 @@
 package evm
 
 import (
-	"fmt"
-	"strings"
-
-	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 
@@ -54,34 +50,4 @@ func toGethSignature(s types.Signature) bindings.ManyChainMultiSigSignature {
 		S: [32]byte(s.S.Bytes()),
 		V: s.V,
 	}
-}
-
-// ABIEncode is the equivalent of abi.encode.
-// See a full set of examples https://github.com/ethereum/go-ethereum/blob/420b78659bef661a83c5c442121b13f13288c09f/accounts/abi/packing_test.go#L31
-func ABIEncode(abiStr string, values ...any) ([]byte, error) {
-	// Create a dummy method with arguments
-	inDef := fmt.Sprintf(`[{ "name" : "method", "type": "function", "inputs": %s}]`, abiStr)
-	inAbi, err := abi.JSON(strings.NewReader(inDef))
-	if err != nil {
-		return nil, err
-	}
-
-	res, err := inAbi.Pack("method", values...)
-	if err != nil {
-		return nil, err
-	}
-
-	return res[4:], nil
-}
-
-// abiDecode is the equivalent of abi.decode.
-// See a full set of examples https://github.com/ethereum/go-ethereum/blob/420b78659bef661a83c5c442121b13f13288c09f/accounts/abi/packing_test.go#L31
-func abiDecode(abiStr string, data []byte) ([]any, error) {
-	inDef := fmt.Sprintf(`[{ "name" : "method", "type": "function", "outputs": %s}]`, abiStr)
-	inAbi, err := abi.JSON(strings.NewReader(inDef))
-	if err != nil {
-		return nil, err
-	}
-
-	return inAbi.Unpack("method", data)
 }
